Moved build file name functionality out of DanteClasses and into DanteMain and dante.dpr.
A search for a build file is done only if the -buildfile option was not used. This behavior is consistent with Ant.

The default build file name is AppName.xml, where AppName is the name of the .exe used to invoke the tool. This lets users that have added custom tasks to the tool use a different file name just by renaming the .exe.

If the search for the AppName.xml fails, then the tool searches for dante.xml. If that fails, then the tool searches for build.xml (this is for backwards compatibility).


git-svn-id: svn://svn.code.sf.net/p/want/code/trunk@247 0fbc77ab-9432-0410-b702-c5701067c881