Delete task now has a DeleteReadOnly property allowing read-only files to be deleted if set to true.
I also removed the requirement for the Delete task to have either the file or dir property set. If I want to delete all files in a directory, I can include a FileSet and don't want to delete a specific file or directory in that case. The task was already coded to actually do this work, but the Init wouldn't allow it to be used this way.


git-svn-id: svn://svn.code.sf.net/p/want/code/trunk@268 0fbc77ab-9432-0410-b702-c5701067c881